Title: 최솟값 3개, Time: 347ms, Memory: 21MB, Status: Passed - CodeTree

+++ 240501/최솟값 3개/top-3-smallest-number.java
@@ -11,7 +11,7 @@ public class Main {
         int n = Integer.parseInt(br.readLine());
         st = new StringTokenizer(br.readLine());
         StringBuilder sb = new StringBuilder();
-        int res = 1;
+        long res = 1;
         for (int i = 0; i < n; i++) {
             int num = Integer.parseInt(st.nextToken());
